requirements.txt
Flask==3.0.3
flask-cors==4.0.1
groq==0.11.0
langchain==0.3.0
langchain-community==0.3.0
langchain-huggingface==0.0.3
sentence-transformers==3.1.1
python-dotenv==1.0.1
gunicorn==23.0.0